+++ CHANGELOG.md
@@ -6,7 +6,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
-
+### Added
+- Added support for passing in the granularity when running queries
+- Added fetching the datasets from the backend when prompting to select a dataset
 
 ## [0.0.74] 2023-08-15
 ### Removed